make cph search the first page users interact with
