+++ src/server/home/index.js
@@ -1,7 +1,32 @@
-import { homeController } from './controller.js'
+import Joi from 'joi'
+import { cphSearchController } from '../cph-search/controller.js'
+
+/**
+ * Custom Joi validation for CPH format (XX/XXX/XXXX)
+ */
+const cphSchema = Joi.string().custom((value, helpers) => {
+  // CPH format: XX/XXX/XXXX - exactly 2 digits, slash, 3 digits, slash, 4 digits
+  const parts = value.split('/')
+  if (parts.length !== 3) {
+    return helpers.error('any.invalid')
+  }
+
+  const [first, second, third] = parts
+  if (first.length !== 2 || second.length !== 3 || third.length !== 4) {
+    return helpers.error('any.invalid')
+  }
+
+  // Check each part contains only digits
+  if (!/^\d+$/.test(first) || !/^\d+$/.test(second) || !/^\d+$/.test(third)) {
+    return helpers.error('any.invalid')
+  }
+
+  return value
+}).message('CPH number must be in format XX/XXX/XXXX (e.g. 12/123/1234)')
 
 /**
  * Sets up the routes used in the home page.
+ * Now serves the CPH search functionality as the main landing page.
  * These routes are registered in src/server/router.js.
  */
 
@@ -16,7 +41,30 @@ export const home = {
         {
           method: 'GET',
           path: '/',
-          ...homeController
+          options: {
+            validate: {
+              query: Joi.object({
+                cph: cphSchema.optional(),
+                created: Joi.string().valid('true').optional()
+              }),
+              failAction: async (request, h, error) => {
+                const validationError = error.details?.[0]?.message || 'Invalid CPH format'
+                return h.view('cph-search/index', {
+                  pageTitle: 'Search for Holding',
+                  heading: 'Search for Holding',
+                  breadcrumbs: [
+                    {
+                      text: 'Search for Holding'
+                    }
+                  ],
+                  cph: request.query?.cph,
+                  holding: null,
+                  errorMessage: validationError
+                }).takeover()
+              }
+            }
+          },
+          ...cphSearchController
         }
       ])
     }
